Relocated compare_zmats from conformers to converter
It's not really being used anywhere in ARC at the moment (this PR replaces it with compare_confs which is based on dmats)
Also, removed compare_xyz from conformer (almost_equal_coords_lists in common does the same thing)
