+++ arc/species/conformers.py
@@ -1562,71 +1562,6 @@ def generate_monoatomic_conformer(symbol):
     return conf
 
 
-def compare_zmats(z1, z2, r_tol=0.01, a_tol=2, d_tol=2, verbose=False, symmetric_torsions=None, index=1):
-    """
-    Compare internal coordinates of two conformers of the same species.
-    The comparison could principally be done using all dihedrals, which is information this module readily has,
-    but this function uses Z matrices instead for better robustness (this way rings are considered as well).
-
-    Args:
-        z1 (dict): Z matrix of conformer 1.
-        z2 (dict): Z matrix of conformer 2.
-        r_tol (float, optional): A tolerance for comparing distances (in Angstrom).
-        a_tol (float, optional): A tolerance for comparing angles (in degrees).
-        d_tol (float, optional): A tolerance for comparing dihedral angles (in degrees).
-        verbose (bool, optional): Whether to print a reason for determining the zmats are different if they are,
-                                  ``True`` to print.
-        symmetric_torsions (dict, optional): Keys are tuples scan indices (0- or 1-indexed), values are internal
-                                             rotation symmetry numbers (sigma). Conformers which only differ by an
-                                             integer number of 360 degrees / sigma are considered identical.
-        index (int, optional): Either ``0`` or ``1`` to specify the starting index in the keys of ``symmetric_torsions``
-
-    Returns:
-        bool: Whether the coordinates represent the same conformer within the given tolerance, ``True`` if they do.
-
-    Raises:
-        InputError: If ``xyz1`` and ``xyz2`` are of wrong type.
-    """
-    # convert the keys of symmetric_torsions to 0-indexed torsion tuples
-    symmetric_torsions = {tuple([torsion[i] - index for i in range(4)]): sigma
-                          for torsion, sigma in symmetric_torsions.items()} if symmetric_torsions is not None else None
-    if not all(isinstance(z, dict) for z in [z1, z2]):
-        raise InputError(f'xyz1 and xyz2 must be dictionaries, got {type(z1)} and {type(z2)}, respectively')
-    if z1['symbols'] != z2['symbols']:
-        return False
-    return zmat.compare_zmats(z1, z2, r_tol=r_tol, a_tol=a_tol, d_tol=d_tol, verbose=verbose,
-                              symmetric_torsions=symmetric_torsions)
-
-
-def compare_xyz(xyz1, xyz2, precision=0.1):
-    """
-    Compare coordinates of two conformers of the same species. Not checking isotopes.
-    compare_zmats() is much more accurate, use this function only if the conformers do not have the `zmat` attribute.
-
-    Args:
-        xyz1 (dict): Coordinates of conformer 1 in either string or array format.
-        xyz2 (dict): Coordinates of conformer 2 in either string or array format.
-        precision (float, optional): The allowed difference threshold between coordinates, in Angstroms.
-
-    Returns:
-        bool: Whether the coordinates represent the same conformer within the given ``precision``, ``True`` if they do.
-
-    Raises:
-        InputError: If ``xyz1`` and ``xyz2`` are of wrong type or have different elements (not considering isotopes).
-    """
-    if not all(isinstance(xyz, dict) for xyz in [xyz1, xyz2]):
-        raise InputError('xyz1 and xyz2 must be dictionaries, got {0} and {1}, respectively'.format(
-                              type(xyz1), type(xyz2)))
-    if xyz1['symbols'] != xyz2['symbols']:
-        raise IndexError('xyz1 and xyz2 have different elements, cannot compare coordinates. '
-                         'Got:\n{0}\nand:\n{1}'.format(xyz1['symbols'], xyz2['symbols']))
-    for coord1, coord2 in zip(xyz1['coords'], xyz2['coords']):
-        for entry1, entry2 in zip(coord1, coord2):
-            if abs(entry1 - entry2) > precision:
-                return False
-    return True
-
-
 def translate_groups(label, mol, xyz, pivot):
     """
     Exchange between two groups in a molecule. The groups cannot share a ring with the pivotal atom.
